+++ src/city_gen_model/CityStateParameters.java
@@ -4,20 +4,20 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class CityStateParameters {
-    Map<CityParamType, Integer> quantityMap;
+    Map<CityParamType, Double> quantityMap;
 
     public CityStateParameters() {
         quantityMap = new HashMap<>();
         for (CityParamType cityParamType : CityParamType.values()) {
-            quantityMap.put(cityParamType, 0);
+            quantityMap.put(cityParamType, 0.0);
         }
     }
 
-    public int getQuantity(CityParamType paramType) {
+    public double getQuantity(CityParamType paramType) {
         return quantityMap.get(paramType);
     }
 
-    public void setQuantity(CityParamType paramType, int value) {
+    public void setQuantity(CityParamType paramType, double value) {
         quantityMap.put(paramType, value);
     }
 }
