+++ src/city_gen_model/progression/ExponentialFunction.java
@@ -9,7 +9,7 @@ public class ExponentialFunction implements Progression{
     }
 
     @Override
-    public double getNextValue(double currentValue, int timeInterval) {
+    public double getNextValue(double currentValue, double timeInterval) {
         return Math.max(currentValue * Math.pow(rateOfChange, timeInterval), 0);
     }
 }
